requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
cryptography==43.0.3
Flask==3.0.3
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
gevent==24.10.3
geventhttpclient==2.3.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.32.0
MarkupSafe==3.0.2
msgpack==1.1.0
prometheus_client==0.21.0
prometheus_flask_exporter==0.23.1
psutil==6.1.0
pycparser==2.22
PyJWT==2.9.0
pyzmq==26.2.0
requests==2.32.3
setuptools==75.2.0
urllib3==2.2.3
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.1.0
